# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Reverse Match API
  version: 1.0.0
  summary: REST contract between the Reverse Match Flutter client and the Node/Express backend.
  description: |
    Single source of truth for the Reverse Match HTTP API. Mirrors the contract previously
    documented in `BACKEND_API.md`. Socket.IO events live in the sibling `asyncapi.yaml`
    document.

    Conventions:
      - Base path: `/api/v1`
      - Auth: `Authorization: Bearer <accessToken>` on every route except `/auth/*` and `/config`.
      - Pagination: cursor (`?cursor=`) for the swipe feed; page-based (`?page=&limit=`) for matches and messages.
      - Errors: all non-2xx responses share the shape defined by `ErrorResponse`.
      - IDs: MongoDB `_id` (24-char hex) serialized as a string. Both `id` and `_id` are accepted by the client; `_id` is preferred.
      - Dates: ISO-8601 UTC strings (e.g. `2026-05-02T14:30:00.000Z`).
      - Coordinates: GeoJSON order — `[longitude, latitude]`.
  license:
    name: Proprietary
  contact:
    name: Reverse Match Team
    email: support@reversematch.app
servers:
  - url: http://localhost:5000/api/v1
    description: Local development
  - url: https://api.reversematch.app/api/v1
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /boost/status:
    get:
      tags: [boost]
      summary: Current user's active boost (if any)
      operationId: getBoostStatus
      responses:
        '200':
          description: Boost status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoostStatusModel'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /boost/purchase:
    post:
      tags: [boost]
      summary: Initiate a Stripe Checkout (or IAP) purchase for a boost tier
      description: |
        Returns a checkout URL. On `url: ""` the frontend silently skips the launch
        (used for free-tier fallbacks). After successful payment, backend activates the
        boost via Stripe webhook.
      operationId: purchaseBoost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoostPurchaseRequest'
      responses:
        '200':
          description: Checkout URL issued
          content:
            application/json:
              schema:
                type: object
                required: [url]
                properties:
                  url:
                    type: string
                    description: External checkout URL. Empty string disables the redirect.
                    examples:
                      - https://checkout.stripe.com/c/pay/cs_test_...
                      - ""
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'

  # ----------------------------------------------------------------------
  # 8. Safety
  # ----------------------------------------------------------------------
...
  /account:
    delete:
      tags: [account]
      summary: Permanently delete the authenticated user's account
      description: |
        Cascades: user row, photos (Cloudinary), likes (sent + received), matches, messages,
        boost records, refresh tokens, FCM tokens. Requires the magic confirmation string
        to prevent accidents.
      operationId: deleteAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountDeleteRequest'
      responses:
        '200':
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyObject'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /account/export:
    get:
      tags: [account]
      summary: Export all data the authenticated user has stored on the platform
      description: |
        GDPR Article 15 — right of access. Returns a JSON document of the
        user's profile, photos (URLs), matches, messages they sent, likes
        sent + received, and current boost state. Response is delivered as
        `Content-Disposition: attachment` so a browser download triggers a
        `.json` file save; an in-app fetch sees a regular JSON body.
      operationId: exportAccountData
      responses:
        '200':
          description: Exported data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountExport'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'

  # ----------------------------------------------------------------------
  # 9b. Admin — selfie review queue (Phase 1.3)
  # ----------------------------------------------------------------------
  /admin/selfies/pending:
    get:
      tags: [admin]
      summary: List users whose selfie is awaiting admin review
      description: |
        Page-paginated list of users with `selfieReviewStatus = 'pending'`.
        Sorted by `updatedAt DESC` so the most recent uploads surface first.
      operationId: listPendingSelfies
      parameters:
        - $ref: '#/components/parameters/Page'
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Pending selfies page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingSelfiesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/selfies/{userId}/approve:
    post:
      tags: [admin]
      summary: Approve a user's selfie (grants the verified badge)
      operationId: approveSelfie
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            $ref: '#/components/schemas/ObjectId'
      responses:
        '200':
          description: Selfie approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSelfieReviewResult'
        '400':
          description: Selfie is not in `pending` state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/selfies/{userId}/reject:
    post:
      tags: [admin]
      summary: Reject a user's selfie (with optional reason)
      operationId: rejectSelfie
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            $ref: '#/components/schemas/ObjectId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSelfieRejectRequest'
      responses:
        '200':
          description: Selfie rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSelfieReviewResult'
        '400':
          description: Selfie is not in `pending` state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

  # ----------------------------------------------------------------------
  # 10. Config (unauthenticated)
  # ----------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `Authorization: Bearer <accessToken>` on every request except `/auth/*` and `/config`.
        On HTTP 401 the client calls `POST /auth/refresh-token` once and retries the original
        request. If refresh fails the user is logged out client-side.
  parameters:
    Page:
      in: query
      name: page
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: 1-indexed page number.

    # -------- Phase 1.4 -----------------------------------------------
...
  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing, invalid, or expired access token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but not allowed to access this resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ObjectId:
      type: string
      pattern: '^[a-fA-F0-9]{24}$'
      description: 24-character MongoDB ObjectId hex string.
      examples:
        - 64f5b7c3e4d6f8a1b2c3d4e5
    IsoDateTime:
      type: string
      format: date-time
      description: ISO-8601 UTC timestamp.
      examples:
        - 2026-05-02T14:30:00.000Z
    EmptyObject:
      type: object
      description: An empty JSON object — `{}`.
      additionalProperties: false
...
    Gender:
      type: string
      enum: [male, female, nonbinary]
      description: Self-identified gender.
    GenderPreference:
      type: string
      enum: [men, women, everyone]
      description: Who the user wants to see in their feed.
    DatingIntention:
      type: string
      enum:
        - life_partner
        - long_term
        - long_term_open_short
        - short_term_open_long
        - short_term
        - new_friends
        - figuring_out
    RelationshipType:
      type: string
      enum: [monogamy, non_monogamy, open_to_exploring, prefer_not_to_say]
    Exercise:
      type: string
      enum: [active, sometimes, never, prefer_not_to_say]
    Zodiac:
      type: string
      enum:
        - aries
        - taurus
        - gemini
        - cancer
        - leo
        - virgo
        - libra
        - scorpio
        - sagittarius
        - capricorn
        - aquarius
        - pisces
    Children:
      type: string
      enum: [have, dont_have, prefer_not_to_say]
    FamilyPlans:
      type: string
      enum: [want, dont_want, open, not_sure, prefer_not_to_say]
    ViceFrequency:
      type: string
      enum: [yes, sometimes, rarely, no, prefer_not_to_say]
      description: How often the user engages in a given vice.
    Politics:
      type: string
      enum: [liberal, moderate, conservative, not_political, prefer_not_to_say]
      description: |
        Self-reported political leaning. Consumed by `idealMatch.service` scoring; must
        be one of the listed values (no free-form string).
    BoostLevel:
      type: string
      enum: [none, bronze, silver, gold]
    BoostTier:
      type: string
      enum: [bronze, silver, gold]
    LikeStatus:
      type: string
      enum: [pending, accepted, rejected]
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable error code (e.g. `VALIDATION_FAILED`, `OTP_EXPIRED`).
              examples:
                - VALIDATION_FAILED
            message:
              type: string
              description: Human-readable message — the frontend surfaces this in a snackbar.
              examples:
                - Email is required
            requestId:
              type: string
              description: |
                Correlation ID from the request-id middleware (mirrors the `X-Request-Id`
                response header). Optional — present on responses served after the requestId
                middleware mounted; absent on older fixtures or pre-middleware errors.

    # -------- Auth payloads --------
...
    AccountDeleteRequest:
      type: object
      required: [confirmation]
      properties:
        confirmation:
          type: string
          enum: [DELETE_MY_ACCOUNT]
          description: Magic string the client sends to acknowledge intent.
    BoostPurchaseRequest:
      type: object
      required: [tier]
      properties:
        tier:
          $ref: '#/components/schemas/BoostTier'

    # -------- Domain objects --------
    Photo:
      type: object
      required: [url, publicId]
      properties:
        url:
          type: string
          format: uri
          description: Public Cloudinary delivery URL.
          examples:
            - https://res.cloudinary.com/reverse-match/image/upload/v123/users/u123/abc.jpg
        publicId:
          type: string
          description: Cloudinary `publicId` — `<userId>/<random>` convention. Used for deletes and reorders.
          examples:
            - users/u123/abc
    Prompt:
      type: object
      required: [question, answer]
      properties:
        question:
          type: string
          examples:
            - My simple pleasures
        answer:
          type: string
          maxLength: 300
          examples:
            - Filter coffee on a rainy morning
    Vices:
      type: object
      description: Each vice is independently set; all keys optional.
      properties:
        drinking:
          $ref: '#/components/schemas/ViceFrequency'
        smoking:
          $ref: '#/components/schemas/ViceFrequency'
        marijuana:
          $ref: '#/components/schemas/ViceFrequency'
        drugs:
          $ref: '#/components/schemas/ViceFrequency'
      additionalProperties: false
    Location:
      type: object
      description: |
        Stored as GeoJSON Point in Mongo with a `2dsphere` index on `coordinates`.
      required: [coordinates]
      properties:
        coordinates:
          type: array
          description: |
            GeoJSON coordinate pair — `[longitude, latitude]`, NOT `[latitude, longitude]`.
            Backend enforces order on save; frontend constructs it from the device geolocator.
          minItems: 2
          maxItems: 2
          items:
            type: number
          examples:
            - [72.8777, 19.0760]
        city:
          type: string
          examples:
            - Mumbai
        state:
          type: string
          examples:
            - Maharashtra
    Preferences:
      type: object
      description: Match filtering preferences.
      properties:
        ageMin:
          type: integer
          minimum: 18
          maximum: 100
          default: 18
        ageMax:
          type: integer
          minimum: 18
          maximum: 100
          default: 99
        maxDistance:
          type: integer
          minimum: 1
          description: Max distance in kilometers.
          examples:
            - 50
        genderPreference:
          $ref: '#/components/schemas/GenderPreference'
    UserModel:
      type: object
      description: |
        Canonical user shape returned by the backend. Fields the frontend defaults when missing:
        - `pronouns`, `orientation`, `interests`, `photos`, `prompts`, `ethnicity`, `languages` → `[]`
        - `daysWithoutMatch` → `0`
        - `boostLevel` → `"none"`
        - `isProfileComplete`, `isVerified` → `false`
      required: [_id]
      properties:
        _id:
          $ref: '#/components/schemas/ObjectId'
        email:
          type: string
          format: email
        name:
          type: string
        age:
          type: integer
          minimum: 18
          maximum: 100
          description: Server-computed from `dob`. Frontend may also derive on render.
        dob:
          $ref: '#/components/schemas/IsoDateTime'
        gender:
          $ref: '#/components/schemas/Gender'
        pronouns:
          type: array
          items:
            type: string
          default: []
        orientation:
          type: array
          items:
            type: string
          default: []
        bio:
          type: string
          maxLength: 300
        exercise:
          $ref: '#/components/schemas/Exercise'
        zodiac:
          $ref: '#/components/schemas/Zodiac'
        interests:
          type: array
          items:
            type: string
          default: []
        photos:
          type: array
          items:
            $ref: '#/components/schemas/Photo'
          minItems: 0
          maxItems: 6
          default: []
        prompts:
          type: array
          items:
            $ref: '#/components/schemas/Prompt'
          default: []
        height:
          type: integer
          minimum: 100
          maximum: 230
          description: Height in cm.
        ethnicity:
          type: array
          items:
            type: string
          default: []
        children:
          $ref: '#/components/schemas/Children'
        familyPlans:
          $ref: '#/components/schemas/FamilyPlans'
        hometown:
          type: string
        jobTitle:
          type: string
        workplace:
          type: string
        education:
          type: string
        religion:
          type: string
        politics:
          $ref: '#/components/schemas/Politics'
        languages:
          type: array
          items:
            type: string
          default: []
        datingIntentions:
          $ref: '#/components/schemas/DatingIntention'
        relationshipType:
          $ref: '#/components/schemas/RelationshipType'
        vices:
          $ref: '#/components/schemas/Vices'
        location:
          $ref: '#/components/schemas/Location'
        preferences:
          $ref: '#/components/schemas/Preferences'
        daysWithoutMatch:
          type: integer
          minimum: 0
          default: 0
        boostLevel:
          $ref: '#/components/schemas/BoostLevel'
          default: none
        isProfileComplete:
          type: boolean
          default: false
        isVerified:
          type: boolean
          default: false
        createdAt:
          $ref: '#/components/schemas/IsoDateTime'
...
    BoostStatusModel:
      type: object
      required: [active]
      properties:
        active:
          type: boolean
        tier:
          oneOf:
            - $ref: '#/components/schemas/BoostTier'
            - type: 'null'
        expiresAt:
          oneOf:
            - $ref: '#/components/schemas/IsoDateTime'
            - type: 'null'

    # -------- List wrappers --------
...
    PendingSelfieUser:
      type: object
      description: |
        Trimmed projection of a user awaiting selfie review. Includes the
        selfie itself (URL + publicId) so the admin UI can render it; full
        profile fields are NOT included to keep the queue payload small.
      required: [_id, photos, selfiePhoto, selfieReviewStatus]
      properties:
        _id:
          $ref: '#/components/schemas/ObjectId'
        name:
          type: string
        email:
          type: string
          format: email
        photos:
          type: array
          items:
            $ref: '#/components/schemas/Photo'
        selfiePhoto:
          $ref: '#/components/schemas/Photo'
        isVerified:
          type: boolean
        selfieReviewStatus:
          type: string
          enum: [none, pending, approved, rejected]
        updatedAt:
          $ref: '#/components/schemas/IsoDateTime'
    PendingSelfiesResponse:
      type: object
      required: [users, page, limit, total, hasMore]
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/PendingSelfieUser'
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        hasMore:
          type: boolean
    AdminSelfieRejectRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 200
          description: Optional free-text reason. Surfaced to the user via FCM push.
    AdminSelfieReviewResult:
      type: object
      required: [userId, selfieReviewStatus]
      properties:
        userId:
          $ref: '#/components/schemas/ObjectId'
        isVerified:
          type: boolean
          description: Present on approval only.
        selfieReviewStatus:
          type: string
          enum: [approved, rejected]

    # -------- Phase 1.17 — GDPR export --------
    AccountExportPhoto:
      type: object
      required: [url, publicId]
      properties:
        url:
          type: string
          format: uri
        publicId:
          type: string
    AccountExportMatch:
      type: object
      required: [_id, users, createdAt]
      properties:
        _id:
          $ref: '#/components/schemas/ObjectId'
        users:
          type: array
          items:
            $ref: '#/components/schemas/ObjectId'
        createdAt:
          $ref: '#/components/schemas/IsoDateTime'
        lastMessageAt:
          oneOf:
            - $ref: '#/components/schemas/IsoDateTime'
            - type: 'null'
    AccountExportMessage:
      type: object
      required: [_id, matchId, text, createdAt]
      properties:
        _id:
          $ref: '#/components/schemas/ObjectId'
        matchId:
          $ref: '#/components/schemas/ObjectId'
        text:
          type: string
        createdAt:
          $ref: '#/components/schemas/IsoDateTime'
    AccountExportLike:
      type: object
      required: [_id, status, createdAt]
      properties:
        _id:
          $ref: '#/components/schemas/ObjectId'
        fromUser:
          $ref: '#/components/schemas/ObjectId'
        toUser:
          $ref: '#/components/schemas/ObjectId'
        status:
          $ref: '#/components/schemas/LikeStatus'
        createdAt:
          $ref: '#/components/schemas/IsoDateTime'
    AccountExport:
      type: object
      required: [exportedAt, schemaVersion, profile, photos, matches, messagesSent, likesSent, likesReceived, boost]
      properties:
        exportedAt:
          $ref: '#/components/schemas/IsoDateTime'
        schemaVersion:
          type: integer
          description: Bump when the export shape changes. v1 = the schema below.
          examples: [1]
        profile:
          $ref: '#/components/schemas/UserModel'
        photos:
          type: array
          items:
            $ref: '#/components/schemas/AccountExportPhoto'
        matches:
          type: array
          items:
            $ref: '#/components/schemas/AccountExportMatch'
        messagesSent:
          type: array
          description: Messages the user sent. Received messages are NOT included (they belong to the other user's right of access).
          items:
            $ref: '#/components/schemas/AccountExportMessage'
        likesSent:
          type: array
          items:
            $ref: '#/components/schemas/AccountExportLike'
        likesReceived:
          type: array
          items:
            $ref: '#/components/schemas/AccountExportLike'
        boost:
          type: object
          properties:
            level:
              $ref: '#/components/schemas/BoostLevel'
            expiry:
              oneOf:
                - $ref: '#/components/schemas/IsoDateTime'
                - type: 'null'